+++ CHANGELOG.md
@@ -12,6 +12,7 @@ for open development tasks and limitations.
 - `binst` sub-cmd for build and installing
 - `clean` sub-cmd for wiping build related folders
 - `qtest` sub-cmd for running black, flake8, and unit-tests
+- better trace / debug log messages
 -